:recycle: certif: combine rejection and validation modals
