+++ certif/app/components/session-supervising/live-alert-handled-modal.hbs
@@ -1,0 +1,32 @@
+<PixModal
+  @class="pix-modal-dialog live-alert-handled-modal"
+  @title={{@title}}
+  @onCloseButtonClick={{@closeConfirmationModal}}
+  @showModal={{@showModal}}
+>
+  <:content>
+    <p>
+      <PixMessage @type="success" class="live-alert-handled-modal__success-banner" @withIcon={{true}}>
+        {{#if @isLiveAlertValidated}}
+          {{t "pages.session-supervising.candidate-in-list.handle-live-alert-modal.handled.validated-banner"}}
+        {{else}}
+          {{t "pages.session-supervising.candidate-in-list.handle-live-alert-modal.handled.rejected-banner"}}
+        {{/if}}
+      </PixMessage>
+    </p>
+
+    <div class="app-modal-body__warning live-alert-handled-modal__description">
+      <p>
+        {{t "pages.session-supervising.candidate-in-list.handle-live-alert-modal.handled.description"}}
+      </p>
+    </div>
+  </:content>
+  <:footer>
+    <PixButton
+      @triggerAction={{@closeConfirmationModal}}
+      aria-label={{t "pages.session-supervising.candidate-in-list.handle-live-alert-modal.handled.close-button-label"}}
+    >
+      {{t "pages.session-supervising.candidate-in-list.handle-live-alert-modal.handled.close-button-text"}}
+    </PixButton>
+  </:footer>
+</PixModal>
